#7 restore the status of all resources
This is necessary to restore backup resources without executing the
backup again.
It also helps to prevent the dogu operator and possibly other operators
from being confused after a restore.
However, we don't know if this has any other implications.
If problems arise, we'll have to give this more granularity.